# docs/openapi.json
{
...
  "components": {
    "schemas": {
...
      "CommandAcceptedDto": {
        "description": "202 body returned by every Workflow Command (DSD-322: silent success\nis forbidden — every accepted command echoes back a small envelope so\nthe caller has something to correlate against the WS broadcast).",
        "properties": {
          "accepted_at": {
            "description": "Wall-clock when the State Update record was appended (DSD-622).",
            "type": "string"
          },
          "command": {
            "description": "Stable command name: `dispatch_next` / `accept_review` / `escalate` /\n`finalize_slice` / `resume_slice`. Mirrors DDD §3.1 command names.",
            "type": "string"
          },
          "project_id": {
            "description": "Project this command targets.",
            "type": "string"
          },
          "request_id": {
            "description": "Request-id correlator (same UUIDv7 returned in `X-Request-Id`).",
            "type": "string"
          },
          "slice_id": {
            "description": "Slice this command targets. `None` for `dispatch_next` because the\nnext slice is chosen inside the lock and broadcast over the WS.",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "command",
          "project_id",
          "request_id",
          "accepted_at"
        ],
        "type": "object"
      },
      "ConfirmEscalateDto": {
        "description": "Response from POST /projects/{id}/slices/{slice_id}/confirm-escalate.",
        "properties": {
          "confirmation_token": {
            "type": "string"
          },
          "expires_at": {
            "description": "Wall-clock when the token expires (DSD-622). Five minutes from mint.",
            "type": "string"
          },
          "slice_id": {
            "type": "string"
          }
        },
        "required": [
          "confirmation_token",
          "slice_id",
          "expires_at"
        ],
        "type": "object"
      },
      "EmptyCommandBody": {
        "description": "Empty body shape accepted by accept / finalize / resume. Reserved for\nfuture optional notes; defining the type now keeps the OpenAPI surface\nstable as fields land.",
        "properties": {
          "note": {
            "description": "Optional operator note recorded with the State Update. Not used by\nthe v1 state machine — present so future audit work has a place\nwithout a wire-shape change.",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "description": "Stable SCREAMING_SNAKE_CASE machine string. Once shipped, never reused.",
            "type": "string"
          },
          "details": {
            "description": "Optional structured detail. Omitted on auth failures per INV-A4.",
            "nullable": true
          },
          "message": {
            "description": "Human-readable, lowercase, no trailing period.",
            "type": "string"
          },
          "request_id": {
            "description": "UUIDv7 request correlator echoed in the `X-Request-Id` header.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "type": "object"
      },
      "EscalateRequest": {
        "description": "Body shape for POST /projects/{id}/slices/{slice_id}/escalate. The\nconfirmation token is minted by `confirm-escalate` (DSD-330).",
        "properties": {
          "confirmation_token": {
            "description": "Token previously returned by `POST .../confirm-escalate` for this\nslice. Single-use, scoped to the slice it was minted against.",
            "type": "string"
          },
          "reason": {
            "description": "Optional human-readable rationale recorded on the State Update.",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "confirmation_token"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Mutagen workflow service. Spec is generated from handler annotations via utoipa per ADR-0002.",
    "license": {
      "name": ""
    },
    "title": "mutagen-service",
    "version": "0.3.3"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/projects/{project_id}/dispatch-next": {
      "post": {
        "operationId": "dispatch_next",
        "parameters": [
          {
            "description": "Project UUIDv7",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              }
            },
            "description": "Command accepted; completion broadcast on WS"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown project_id"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Project lock held by another writer"
          }
        },
        "tags": [
          "workflow"
        ]
      }
    },
...
    "/projects/{project_id}/slices/{slice_id}/accept": {
      "post": {
        "operationId": "accept",
        "parameters": [
          {
            "description": "Project UUIDv7",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Slice id",
            "in": "path",
            "name": "slice_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyCommandBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              }
            },
            "description": "Slice accept_review recorded"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown project_id"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Project lock held by another writer"
          }
        },
        "tags": [
          "workflow"
        ]
      }
    },
    "/projects/{project_id}/slices/{slice_id}/confirm-escalate": {
      "post": {
        "operationId": "confirm_escalate",
        "parameters": [
          {
            "description": "Project UUIDv7",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Slice id",
            "in": "path",
            "name": "slice_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmEscalateDto"
                }
              }
            },
            "description": "Single-use confirmation token for /escalate"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown project_id"
          }
        },
        "tags": [
          "workflow"
        ]
      }
    },
    "/projects/{project_id}/slices/{slice_id}/escalate": {
      "post": {
        "operationId": "escalate",
        "parameters": [
          {
            "description": "Project UUIDv7",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Slice id",
            "in": "path",
            "name": "slice_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscalateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              }
            },
            "description": "Slice escalated; State Update written"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown project_id"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Project lock held by another writer"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or stale confirmation token"
          }
        },
        "tags": [
          "workflow"
        ]
      }
    },
    "/projects/{project_id}/slices/{slice_id}/finalize": {
      "post": {
        "operationId": "finalize",
        "parameters": [
          {
            "description": "Project UUIDv7",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Slice id",
            "in": "path",
            "name": "slice_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyCommandBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              }
            },
            "description": "Slice finalize recorded"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown project_id"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Project lock held by another writer"
          }
        },
        "tags": [
          "workflow"
        ]
      }
    },
    "/projects/{project_id}/slices/{slice_id}/resume": {
      "post": {
        "operationId": "resume",
        "parameters": [
          {
            "description": "Project UUIDv7",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Slice id",
            "in": "path",
            "name": "slice_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyCommandBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              }
            },
            "description": "Slice resume recorded"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unknown project_id"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Project lock held by another writer"
          }
        },
        "tags": [
          "workflow"
        ]
      }
    },
...
  }
}